5-flip_bits.c Function that returns the number of bits you would need to flip to get from one number to another

+++ bit_manipulation/5-flip_bits.c
@@ -17,7 +17,7 @@ unsigned int flip_bits(unsigned long int n, unsigned long int m)
  *
  * Return: The count of set bits in the given integer.
  */
-int countSetBits(int n)
+unsigned long int countSetBits(unsigned long int n)
 {
 	int count = 0;
 	while (n > 0)
